Backend/requirements.txt
asgiref==3.9.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cloudinary==1.44.1
cryptography==46.0.2
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
